requirements.txt
# Core dependencies for local, offline document AI system
# All packages are free and can be used offline

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Text processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Embeddings (sentence-transformers for local models)
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector storage (local options)
faiss-cpu>=1.7.4  # FAISS for efficient similarity search
# chromadb>=0.4.0  # Uncomment for ChromaDB support

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Type hints
typing-extensions>=4.5.0

# Optional: Question Answering
# GROQ API (Recommended - fast, cloud-based)
groq>=0.4.0  # GROQ API for fast question answering

# Local LLM (Alternative - fully offline)
# llama-cpp-python>=0.2.0  # Uncomment to enable local LLM QA
# Note: Download GGUF models from HuggingFace (e.g., Mistral 7B)

# Streamlit Dashboard
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0
pandas>=2.0.0

